requirements.txt
asgiref==3.7.2
asttokens==2.2.1
async-timeout==2.0.1
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
daphne==3.0.2
debugpy==1.6.7
decorator==4.4.2
Django==4.2.3
django-cors-headers==3.12.0
django-crispy-forms==2.0
django-redis==5.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
executing==1.2.0
fonttools==4.39.4
gunicorn==20.1.0
htmlmin==0.1.12
hyperlink==21.0.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
incremental==22.10.0
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
moviepy==1.0.3
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.23.5
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.3
proglog==0.1.10
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.9
pygame==2.5.0
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
redis==4.5.1
requests==2.28.2
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post5
soupsieve==2.4.1
sqlparse==0.4.2
stack-data==0.6.2
tangled-up-in-unicode==0.2.0
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.14
wcwidth==0.2.6
whitenoise==6.1.0
youtube-dl==2021.12.17
zope.interface==5.5.2
